requirements.txt
python-telegram-bot==20.7
anthropic==0.75.0
apscheduler==3.11.2
python-dotenv


